upgrade

+++ NutriBase.Logic/Entities/Base/Grocery.cs
@@ -8,8 +8,8 @@ namespace NutriBase.Logic.Entities.Base;
 [Table("Groceries", Schema = "Base")]
 public class Grocery: Product
 {
-    public int? KaloriesPer100Gram { get; set; }
-    public int? ProteinPer100Gram { get; set; }
-    public int? SugarPer100Gram { get; set; }
+    public double? KaloriesPer100Gram { get; set; }
+    public double? ProteinPer100Gram { get; set; }
+    public double? SugarPer100Gram { get; set; }
     public NutritionForm? NutritionForm { get; set; }
 }
