+++ NutriBase.Logic/Models/Base/GroceryDto.cs
@@ -5,8 +5,8 @@ namespace NutriBase.Logic.Models.Base;
 
 public class GroceryDto : ProductDto
 {
-    public int? KaloriesPer100Gram { get; set; }
-    public int? ProteinPer100Gram { get; set; }
-    public int? SugarPer100Gram { get; set; }
+    public double? KaloriesPer100Gram { get; set; }
+    public double? ProteinPer100Gram { get; set; }
+    public double? SugarPer100Gram { get; set; }
     public NutritionForm? NutritionForm { get; set; }
 }
